adc: add explicit cast when printing _PAGE_SIZE
TASK: MSH-36

+++ adc/ad7779/imx6ull/dma.c
@@ -80,7 +80,7 @@ static int sdma_init(size_t size, size_t count, addr_t *phys_addr)
 	addr_t bd_phys_addr;
 
 	if ((size % _PAGE_SIZE) != 0) {
-		log_error("buffer size is not aligned to %lu", _PAGE_SIZE);
+		log_error("buffer size is not aligned to %zu", (size_t)_PAGE_SIZE);
 		return -1;
 	}
